pyproject.toml
[tool.poetry]
name = "nexhabit-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
fastapi = "^0.115.2"
supabase = "^2.9.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
pytest = "^8.3.3"
psycopg2 = "^2.9.10"
litellm = "^1.50.0"
mangum = "^0.19.0"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
